requirements.txt
streamlit==1.31.1
groq==0.4.2
python-dotenv==1.0.1
